added support for unholy dynamic timing message content...
