+++ modules/ftm/ftmx86/meta.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include "node.h"
-#include "visitor.h"
+
 
 
 class Meta : public Node {
@@ -15,8 +15,8 @@ public:
   Meta(const std::string& name, const uint32_t& hash, uint8_t (&b)[_MEM_BLOCK_SIZE], uint32_t flags) : Node(name, hash, b, flags) {}
   ~Meta()  {};
   virtual void accept(const VisitorVertexWriter& v)     const = 0;
-  virtual void accept(const VisitorNodeUploadCrawler& v)      const = 0;
-  virtual void accept(const VisitorNodeDownloadCrawler& v)    const = 0;
+  virtual void accept(const VisitorUploadCrawler& v)      const = 0;
+  virtual void accept(const VisitorDownloadCrawler& v)    const = 0;
   virtual void show(void)                               const = 0;
   virtual void show(uint32_t cnt, const char* sPrefix)  const = 0;
   virtual void serialise(const vAdr &va) const;
@@ -43,8 +43,8 @@ public:
   : Meta(name, hash, b, ((flags & ~NFLG_TYPE_SMSK) | (NODE_TYPE_QUEUE << NFLG_TYPE_POS))) {}
   ~CmdQMeta()  {};
   virtual void accept(const VisitorVertexWriter& v)         const override { v.visit(*this); }
-  virtual void accept(const VisitorNodeUploadCrawler& v)    const override { v.visit(*this); }
-  virtual void accept(const VisitorNodeDownloadCrawler& v)  const override { v.visit(*this); }
+  virtual void accept(const VisitorUploadCrawler& v)    const override { v.visit(*this); }
+  virtual void accept(const VisitorDownloadCrawler& v)  const override { v.visit(*this); }
 
   void show(void)       const;
   void show(uint32_t cnt, const char* sPrefix) const;
@@ -61,8 +61,8 @@ public:
   : Meta(name, hash, b, ((flags & ~NFLG_TYPE_SMSK) | (NODE_TYPE_QBUF << NFLG_TYPE_POS))) {}
   ~CmdQBuffer()  {};
   virtual void accept(const VisitorVertexWriter& v)         const override { v.visit(*this); }
-  virtual void accept(const VisitorNodeUploadCrawler& v)    const override { v.visit(*this); }
-  virtual void accept(const VisitorNodeDownloadCrawler& v)  const override { v.visit(*this); }
+  virtual void accept(const VisitorUploadCrawler& v)    const override { v.visit(*this); }
+  virtual void accept(const VisitorDownloadCrawler& v)  const override { v.visit(*this); }
 
   void show(void)       const;
   void show(uint32_t cnt, const char* sPrefix)  const;
@@ -78,8 +78,8 @@ public:
   : Meta(name, hash, b, ((flags & ~NFLG_TYPE_SMSK) | (NODE_TYPE_ALTDST << NFLG_TYPE_POS))) {}
   ~DestList()  {};
   virtual void accept(const VisitorVertexWriter& v)         const override { v.visit(*this); }
-  virtual void accept(const VisitorNodeUploadCrawler& v)    const override { v.visit(*this); }
-  virtual void accept(const VisitorNodeDownloadCrawler& v)  const override { v.visit(*this); }
+  virtual void accept(const VisitorUploadCrawler& v)    const override { v.visit(*this); }
+  virtual void accept(const VisitorDownloadCrawler& v)  const override { v.visit(*this); }
 
   void show(void)       const;
   void show(uint32_t cnt, const char* sPrefix)  const;
