+++ modules/ftm/ftmx86/node.h
@@ -4,9 +4,10 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include "common.h"
-#include "visitor.h"
+#include "visitoruploadcrawler.h"
+#include "visitordownloadcrawler.h"
+#include "visitorvertexwriter.h"
 #include "ftm_common.h"
-
 
 
 
@@ -34,14 +35,16 @@ public:
   const std::string&  getName() const {return this->name;}
   const uint32_t&     getHash() const {return this->hash;}
   const uint32_t&     getFlags() const {return this->flags;}
-  const bool isPainted() const {return (bool)(this->flags >> NFLG_PAINT_LM32_POS) & NFLG_PAINT_LM32_MSK;}
+  void     setFlags(uint32_t flags) {this->flags |= flags;}
+  void     clrFlags(uint32_t flags) {this->flags &= ~flags;}
+  const bool isPainted() const {return (bool)(this->flags & NFLG_PAINT_LM32_SMSK);}
   const auto getB() -> uint8_t (&)[_MEM_BLOCK_SIZE] {return this->b;}
 
   virtual void show(void)                               const = 0;
   virtual void show(uint32_t cnt, const char* sPrefix)  const = 0;
   virtual void accept(const VisitorVertexWriter& v)     const = 0;
-  virtual void accept(const VisitorNodeUploadCrawler& v)      const = 0;
-  virtual void accept(const VisitorNodeDownloadCrawler& v)    const = 0;
+  virtual void accept(const VisitorUploadCrawler& v)      const = 0;
+  virtual void accept(const VisitorDownloadCrawler& v)    const = 0;
   virtual bool isMeta(void) const = 0;
   virtual void deserialise() = 0;
   virtual void serialise(const vAdr &va) const {
